+++ src/main.rs
@@ -8,8 +8,8 @@ mod backend;
 use backend::currency::*;
 
 mod modules;
-use modules::saybeans::*;
 use modules::currency::*;
+use modules::showmebeans::*;
 
 use serenity::{
     async_trait,
@@ -29,8 +29,8 @@ struct CommandHandler;
 #[async_trait]
 impl EventHandler for CommandHandler{
     async fn ready(&self, ctx: Context, _data_about_bot: Ready){
-        ctx.set_activity(Activity::competing("Bean eating")).await;
         create_wallet_table();
+        ctx.set_activity(Activity::listening("Quilla - Beans Beans Beans")).await;
         println!("Hello! I am ready to dispatch beans!");
     }
 
@@ -71,8 +71,8 @@ async fn main(){
         .prefix("!")
         .case_insensitivity(true)
         .on_mention(Some(bot_id)))
-        .group(&SAYBEANS_GROUP)
-        .group(&CURRENCY_GROUP);
+        .group(&CURRENCY_GROUP)
+        .group(&SHOWMEBEANS_GROUP);
 
     let mut client = Client::builder(&token)
         .framework(framework)