+++ src/modules/mod.rs
@@ -1,2 +1,2 @@
-pub mod saybeans;
 pub mod currency;
+pub mod showmebeans;