+++ CHANGELOG.md
@@ -3,4 +3,4 @@
 ## Version 0.0.1 *Sprout*
 * Initial version with basic functionality.
 * Basic project structure with serenity framework.
-* **Saybeans** command which prints beans in the chat when issued.
+* **beans** command which prints beans in the chat when issued.